+++ Scene.h
@@ -4,7 +4,7 @@
 class Scene {
 public:
 	virtual ~Scene() {};
-	virtual void Update(double deltaTime) = 0;
+	virtual void Update(float deltaTime) = 0;
 	virtual void Render() = 0;
 	virtual void HandleEvents(GLFWwindow* window) = 0;
